requirements.txt
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7
notion-client==2.2.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0 